### 2021-08-04 #3566 @whedon
[PRE REVIEW]: eyecite: A tool for parsing legal citations
**Submitting author:** @mattdahl (<a href="http://orcid.org/0000-0001-8437-1455">Matthew Dahl</a>)
**Repository:** <a href="https://github.com/freelawproject/eyecite" target ="_blank">https://github.com/freelawproject/eyecite</a>
**Version:** v2.2.0
