### 2021-08-05 #3566 @mattdahl
Please be advised that the paper is not in our main branch; you can find it here instead: https://github.com/freelawproject/eyecite/tree/joss-paper/joss
