### 2021-08-06 #3566 @Kevin-Mattheus-Moerman
@danielskatz given your work on citations for software / software for citations, would this be something you could help edit once you are available?
