### 2021-08-06 #3566 @danielskatz
👋 @mattdahl - Can you suggest some potential reviewers by mentioning them here in this thread, ideally by github username if known (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
