### 2021-08-12 #3566 @mattdahl
Hi @danielskatz, here are a couple of people from that spreadsheet that I think might be suitable based of their listed areas of expertise:

- step21 (`law, computable law,  language design, logic, knowledge management, document management, general data gathering and analysis`)
- cmaimone (Python, `Social science, text analysis, general data manipulation, data visualization`)
- pmyteh (Python, `Computational social science, especially focused towards communication and political science, and particularly packages focused on text analysis`)
- cjbarrie (Python, `Social science, APIs, text analysis`)
- rlskoeser (Python, `web applications, data modeling and munging, text indexing and parsing, digital humanities`)

Please let us know if you'd like us to generate more candidates. Thanks!
