### 2021-08-13 #3566 @danielskatz
👋 @pmyteh & @cjbarrie - would either or both of you be available to review this JOSS submission?
