### 2021-08-16 #3617 @whedon
[REVIEW]: eyecite: A tool for parsing legal citations
**Submitting author:** @mattdahl (<a href="http://orcid.org/0000-0001-8437-1455">Matthew Dahl</a>)
**Repository:** <a href="https://github.com/freelawproject/eyecite" target ="_blank">https://github.com/freelawproject/eyecite</a>
**Version:** v2.3.0
**Archive:** <a href="https://doi.org/10.7910/DVN/STVKTE" target="_blank">10.7910/DVN/STVKTE</a>
