### 2021-08-19 freelawproject/eyecite#86 @step21
Docs should say that hyperscan is currently linux only, not only x86_64 only.
During my review for JOSS, I noticed that hyperscan seems to be unwilling to compile on anything non-linux.
This is fine, but in the eyecite readme at least it only says 'x86' which normally means it should at least run on macos x86_64 or/and Windows x86_64, so I think this should be clarified.
openjournals/joss-reviews#3617
