### 2021-08-19 #3617 @step21
@danielskatz I confirmed basic functionality so far. The paper does say 'high performance' extraction but otherwise does not make any specific claims. Thus, I am not sure how to evaluate performance or functionality at scale as I currently do not have a dataset ready for this. Though at least I would look at their tests, which is when I happened upon the mentioned issue. I can later do this on a linux machine.
