### 2021-08-19 #3617 @danielskatz
@step21 - as we say in the checklist, if there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.) I think it's therefor up to you to decide if you are ok with a vague "high performance" statement, to ask the submitter to explain, or to disagree.
