### 2021-08-19 freelawproject/eyecite#86 @jcushman
Thanks! I agree it should be documented that the python-hyperscan bindings only officially support linux.

Note that in practice HyperscanTokenizer does currently work on MacOS:

```
❯ python3 --version
Python 3.9.6

❯ brew install hyperscan
...

❯ python3 -m pip install hyperscan eyecite
...

❯ python3 -c "from eyecite import get_citations; from eyecite.tokenizers import HyperscanTokenizer; tokenizer = HyperscanTokenizer(); print(get_citations('1 U.S. 1', tokenizer=tokenizer))"
[FullCaseCitation('1 U.S. 1', groups={'volume': '1', 'reporter': 'U.S.', 'page': '1'}, metadata=FullCaseCitation.Metadata(parenthetical=None, pin_cite=None, year=None, court='scotus', plaintiff=None, defendant=None, extra=None))]
```

The underlying hyperscan C++ library is [cross platform](https://intel.github.io/hyperscan/dev-reference/getting_started.html#software), so it should be possible to get python-hyperscan to compile on other platforms as well.

But this doesn't seem to be a tested configuration for python-hyperscan, so it totally makes sense to document what's officially supported.
