### 2021-08-19 freelawproject/eyecite#86 @step21
Thanks. I didn't try brew, I tried `pip install` which complained about some library so maybe manually installing hyperscan headers would also help. Btw, if you are looking for way to build/test on different architectures I can recommend conda-forge.
