### 2021-08-19 bb741e0
Document hyperscan installation requirements
