### 2021-09-08 #3617 @danielskatz
👋 @cmaimone, @step21 - we're now about 3 weeks into the process. How is it going? I see some progress by @step21, but no checked items from @cmaimone
