### 2021-09-08 #3617 @cmaimone
On my to-do list for next week

On Wed, Sep 8, 2021 at 3:19 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @cmaimone <https://github.com/cmaimone>, @step21
> <https://github.com/step21> - we're now about 3 weeks into the process.
> How is it going? I see some progress by @step21
> <https://github.com/step21>, but no checked items from @cmaimone
> <https://github.com/cmaimone>
>
