+++ pyproject.toml
@@ -29,12 +29,12 @@ include = ["eyecite/py.typed"]
 
 [tool.poetry.dependencies]
 python = "^3.7"
-reporters-db = "^3.2.1"
+reporters-db = "^3.2.3"
 lxml = "^4.6.3"
 pyahocorasick = ">= 1.2"
 diff_match_patch_python = "^1.0.2"
 regex = ">=2020.1.8"
-courts-db = "^0.9.13"
+courts-db = "^0.9.23"
 
 [tool.poetry.dev-dependencies]
 black = "^20.8b1"
