### 2021-09-16 freelawproject/eyecite#90 @cmaimone
Project does not have community guidelines
RE: https://github.com/openjournals/joss-reviews/issues/3617

I don't see any of the expected community guidelines:
