### 2021-09-16 freelawproject/eyecite#91 @cmaimone
Full Text Examples and Tests
RE: https://github.com/openjournals/joss-reviews/issues/3617

I'd like to see both usage examples in the tutorial and tests that use longer text documents -- full cases or other documents similar to what you expect to be parsed.  The examples and the tests seem to use single-line snippets with few exceptions, which I don't think reflect actual expected usage.

Longer examples would help people see how to use the software.  Would also be nice to have some example texts that people can run to get started with the software.

Longer text for the tests could reveal issues that other tests don't -- for example what is the full expected set of citations out of a document.  I think the software also does some co-referencing, yes?  So to also see the tests for that in some of the more complicated cases
