### 2021-09-16 08e876e
feat(docs): Adds contribution guidelines
Fixes: #90
