### 2021-09-16 freelawproject/eyecite#90 @mlissner
Thanks @cmaimone. I added a new section to the readme with some details. Please let me know if you have further concerns and I'll be happy to try to help.
