### 2021-09-16 freelawproject/eyecite#92 @cmaimone
API/Reference Documentation
RE: https://github.com/openjournals/joss-reviews/issues/3617

I see the tutorial in the repo readme, but is there reference/API documentation somewhere -- that lists the functions, classes, etc. in the package, with parameters, etc.  (the type generally generated from docstrings)

I expect this type of documentation for packages (or alternatively much more extensive usage guides)
