### 2021-09-16 freelawproject/eyecite#92 @mlissner
We don't have that at present. I've never generated that. @jcushman do you have any experience with generating this kind of thing?
