### 2021-09-16 freelawproject/eyecite#93 @cmaimone
Performance evaluation?
RE: https://github.com/openjournals/joss-reviews/issues/3617

There aren't performance claims in the paper, but I think maybe there should be -- in the paper or in the tutorial.  If I'm evaluating whether to use this package, I have no idea how good it is.  Is there a standard corpus (or can you make one?) where you can run your code and report some basic stats on how many citations you extract, how many errors, etc?  It doesn't even have to be that big of a set of documents -- but some indication of the types of text/documents you've tested against, and how well it does on those.

I get that this package may be the best available for the task, but I have no idea what that means in practice -- can I rely on the results, or does it miss a lot?
