### 2021-09-16 freelawproject/eyecite#94 @cmaimone
More concise list of functionality
RE: https://github.com/openjournals/joss-reviews/issues/3617

It would help in the tutorial/documentation/readme to have a more concise list of what the functionality of the package is -- what are the main things it does.  This could be as simple as a table of contents for the tutorial sections, if all of the functionality is represented by a section.
