### 2021-09-16 freelawproject/eyecite#93 @mlissner
We do say this in the readme:

> eyecite is an open source tool for extracting legal citations from text. It is used, among other things, to annotate millions of legal documents in the collections of CourtListener and the Caselaw Access Project.

It's kind of hard to count errors, because we have fixed pretty much all of the possible citations. I guess we could say something like, "We don't know of any citations that it is unable to parse."
