### 2021-09-16 freelawproject/eyecite#95 @cmaimone
References to "database" in tutorial
RE: https://github.com/openjournals/joss-reviews/issues/3617

There are references to a "database" in the tutorial - I don't know what these mean.  What database?  Are you talking about linking to external legal databases of some type?  Are these just adding URLs?

It also says things like "fetched from our database" -- but what is "our database"?
