### 2021-09-16 freelawproject/eyecite#95 @cmaimone
Related to the above, the entire section

> Now we want to resolve all of the extracted cites into clusters indexed by the resource they refer to, such as a case or statute. We'll use a custom function to resolve a given full cite to its resource, so we can return our own MyCaseModel for citations we recognize. We'll fall back on returning resolve_full_citation() for citations we don't recognize.
>
> For this simplified example, we'll assume we have a database model MyCaseModel so that MyCaseModel.get_by_citation() will return the case referred to by that citation string. In real life this might be a Django model or Elasticsearch lookup. We'll also assume that the same case has the parallel citations "1 U.S. 2" and "1 S. Ct. 2", so MyCaseModel.get_by_citation("1 U.S. 2") returns the same case as MyCaseModel.get_by_citation("1 S. Ct. 2").

I got lost.  I think you're saying you have a hypothetical database with a table model class in it, separate from this package.  This is super confusing for folks who aren't familiar with this type of setup or don't have their own database.  Do you have to use a database for the citation resolution?  All code in the tutorial (with maybe a few exceptions for special use cases, etc.) should be able to run by a potential user.  I can't test the resolve_citations() functionality because the example code uses a database I don't have (and there aren't instructions to set one up.
