### 2021-09-16 freelawproject/eyecite#90 @cmaimone
I saw the license agreement, but the guidelines were asking for more than that (unless I missed some other part).  See https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors for some suggestions at setting up good contributor guidelines
