### 2021-09-16 #3617 @cmaimone
I did an initial review - as you can see above, I added issues to the repo.

Most are around documentation.  I think this project was pulled out as a separate piece from a larger project.  The documentation still seems to reflect that and isn't yet sufficient for an independent user to come and understand and use the package.  I approached this as if I'm a researcher with a collection of legal texts that I want to pull the citations out of.

Also, the community guidelines issue should still be open - see new comment after you closed it.
