### 2021-09-16 freelawproject/eyecite#93 @cmaimone
Yes - even something clearer along those lines.  We've parsed XXX documents in these repositories and reviewed XXX by hand to ensure that eyecite found all citations in these XXX documents.  Just something to let people know the type of performance they can expect
