### 2021-09-17 #3617 @cmaimone
I should also say that the package is impressive - I get a little focused on the checklist and issues sometimes.  But having done similar (but smaller) projects in the past, I appreciate how much effort went into this, and it's great that it's shared for others to use :)
