### 2021-09-17 #3617 @step21
> 👋 @cmaimone, @step21 - we're now about 3 weeks into the process. How is it going? I see some progress by @step21, but no checked items from @cmaimone

I'll get to it today or tomorrow.
