### 2021-09-18 freelawproject/eyecite#92 @step21
This could maybe also be partially automated. There is at least some use of docstrings in the code, these could f.e. be parsed by pdoc, sphinx or a similar documentation tool.
