### 2021-09-18 freelawproject/eyecite#90 @step21
I would also say - as there is a CLA in repo - if you expect people to sign a CLA for contributions, this should definitely be mentioned in the contribution guidelines.
