### 2021-09-18 #3617 @step21
Hi, I also finished my initial review now.
Generally most things are fine and for the majority of the checklist I agree with @cmaimone. Especially that documentation could be more detailed, and example datasets for testing would be nice, especially to verify functionality and performance, which I partly mentioned when I first looked at the repo.
However there are tests in the repo, so I checked 'automated tests', as they can be automated.  Community Guidelines I checked at first, because I thought for such a project with at this stage limited contributors, the amount of info is enough, except for mention of the CLA - as in their repo there is a contributor license agreement, so if people are expected to sign this for contributions, this should be explicitly mentioned in the readme. (there are also free solutions to automate signing)
Regarding the paper - generally it's fine, but I had the feeling that once or twice it switches topcis - f.e. under 'statement of need' it switches from explaining the need, directly to explaining how eyecite itself works. In my opinion this should be its own section.
