### 2021-09-18 23e0c09
Renames annotate() to annotate_citations().
