+++ eyecite/__init__.py
@@ -1,10 +1,10 @@
-from .annotate import annotate
+from .annotate import annotate_citations
 from .find_citations import get_citations
 from .resolve import resolve_citations
 from .utils import clean_text, dump_citations
 
 __all__ = [
-    "annotate",
+    "annotate_citations",
     "get_citations",
     "clean_text",
     "resolve_citations",
