### 2021-09-18 7372d0d
Moves clean_text() to clean module.
