+++ eyecite/__init__.py
@@ -2,12 +2,10 @@ from .annotate import annotate_citations
 from .clean import clean_text
 from .find_citations import get_citations
 from .resolve import resolve_citations
-from .utils import dump_citations
 
 __all__ = [
     "annotate_citations",
     "get_citations",
     "clean_text",
-    "resolve_citations",
-    "dump_citations",
+    "resolve_citations"
 ]
