+++ eyecite/__init__.py
@@ -1,6 +1,6 @@
 from .annotate import annotate_citations
 from .clean import clean_text
-from .find_citations import get_citations
+from .find import get_citations
 from .resolve import resolve_citations
 
 __all__ = [
