### 2021-09-18 d105b66
feat(docs): Generate API documentation using pdoc.
