+++ eyecite/clean.py
@@ -5,8 +5,19 @@ import lxml.html
 
 
 def clean_text(text, steps: Iterable[Union[str, Callable[[str], str]]]) -> str:
-    """Applies each step in order to text, returning the result.
-    Steps may be the names of functions in eyecite.cleaners, or callables.
+    """Given a list of "cleaning" functions, apply each in sequence to a
+    given text string and return the result. Steps may be the names of
+    functions in `eyecite.clean`, or other custom callables. You may wish to
+    use this tool to pre-process your text before feeding it into
+    `eyecite.find.get_citations`, especially if the text was
+    OCR'd from a PDF.
+
+    Args:
+        text: The text to clean.
+        steps: Any `Iterable` (e.g., a list) of cleaning functions to apply.
+
+    Returns:
+        The cleaned text.
     """
     for step in steps:
         if step in cleaners_lookup:
@@ -24,11 +35,14 @@ def clean_text(text, steps: Iterable[Union[str, Callable[[str], str]]]) -> str:
 
 
 def html(html_content: str) -> str:
-    """Given HTML markup, return only text that is visible
-    Adopted from freelawproject/juriscraper/lib/html_utils.py#L163
+    """Given HTML markup, return only text that would be rendered visibly.
+    Adopted from freelawproject/juriscraper/lib/html_utils.py#L163.
 
-    :param html_content: The HTML string
-    :return: Text that is visible
+    Args:
+        html_content: The HTML string.
+
+    Returns:
+        Text that is visible.
     """
     html_tree = lxml.html.fromstring(html_content)
     text = html_tree.xpath(
@@ -42,18 +56,41 @@ def html(html_content: str) -> str:
 
 
 def inline_whitespace(text: str) -> str:
-    """Collapse multiple spaces or tabs into one space character."""
+    """Collapse multiple spaces or tabs within a string into one space
+    character.
+
+    Args:
+        text: The input string.
+
+    Returns:
+        Text with collapsed spaces and tabs.
+    """
     return re.sub(r"[ \t]+", " ", text)
 
 
 def all_whitespace(text: str) -> str:
-    """Collapse multiple whitespace characters into one space character."""
+    """Collapse multiple whitespace characters within a string into one space
+    character.
+
+    Args:
+        text: The input string.
+
+    Returns:
+        Text with collapsed whitespace characters.
+    """
     return re.sub(r"\s+", " ", text)
 
 
 def underscores(text: str) -> str:
     """Remove strings of two or more underscores that are common
-    in text extracted from PDFs."""
+    in text extracted from PDFs.
+
+    Args:
+        text: The input string.
+
+    Returns:
+        Text without underscores.
+    """
     return re.sub(r"__+", "", text)
 
 
