+++ README.rst
@@ -1,7 +1,7 @@
 eyecite
 ==========
 
-eyecite is an open source tool for extracting legal citations from text. It is used, among other things, to annotate millions of legal documents in the collections of `CourtListener <https://www.courtlistener.com/>`_ and the `Caselaw Access Project <https://case.law/>`_.
+eyecite is an open source tool for extracting legal citations from text. It is used, among other things, to process millions of legal documents in the collections of `CourtListener <https://www.courtlistener.com/>`_ and the `Caselaw Access Project <https://case.law/>`_.
 
 eyecite recognizes a wide variety of citations commonly appearing in American legal decisions, including:
 
@@ -12,20 +12,24 @@ eyecite recognizes a wide variety of citations commonly appearing in American le
 * supra: ``Bush, supra, at 100``
 * id.: ``Id., at 101``
 
-You can see all of the citation patterns eyecite looks for over in `reporters_db <https://github.com/freelawproject/reporters-db>`_.
-
 All contributors, corrections, and additions are welcome!
 
-Background
-==========
-This project is the culmination of `years <https://free.law/2012/05/11/building-a-citator-on-courtlistener/>`_ `of <https://free.law/2015/11/30/our-new-citation-finder/>`_ `work <https://free.law/2020/03/05/citation-data-gets-richer/>`_ to build a citator within CourtListener. This project represents the next step in that development: decoupling the parsing logic and exposing it for third-party use as a standalone Python package.
+Functionality
+=============
 
-Since eyecite was factored out from the CourtListener codebase into a standalone package, it has been developed in collaboration with the Caselaw Access Project.
+eyecite offers four core functions:
+
+* `Extraction <https://freelawproject.github.io/eyecite/find.html>`_: Recognize and extract citations from text, using a database that has been trained on over 55 million existing citations (see all of the citation patterns eyecite looks for over in `reporters_db <https://github.com/freelawproject/reporters-db>`_).
+* `Aggregation <https://freelawproject.github.io/eyecite/resolve.html>`_: Aggregate citations with common references (e.g., `supra` and `id.` citations) based on their logical antecedents.
+* `Annotation <https://freelawproject.github.io/eyecite/annotate.html>`_: Annotate citation-laden text with custom markup surrounding each citation, using a fast diffing algorithm.
+* `Cleaning <https://freelawproject.github.io/eyecite/clean.html>`_: Clean and pre-process text for easy use with eyecite.
+
+Read on below for how to get started quickly or for a short tutorial in using eyecite.
 
 Contributions & Support
 =======================
 
-Please see the issues list on Github for things we need, or start a conversation if you have questions or need support.
+Please see the issues list on GitHub for things we need, or start a conversation if you have questions or need support.
 
 If you are fixing bugs or adding features, before you make your first contribution, we'll need a signed contributor license agreement. See the template in the root of the repo for how to get that taken care of.
 
@@ -46,7 +50,7 @@ Install eyecite::
     pip install eyecite
 
 
-Here's a short example of extracting citations and their metadata from text::
+Here's a short example of extracting citations and their metadata from text using eyecite's main :code:`get_citations()` function::
 
     from eyecite import get_citations
 
@@ -84,137 +88,171 @@ Here's a short example of extracting citations and their metadata from text::
 Tutorial
 ==========
 
[…]
