### 2021-09-20 freelawproject/eyecite#98 @step21
Paper structure
Re: https://github.com/openjournals/joss-reviews/issues/3617

Once or twice I had the feeling that the paper switches between topics or mixes things. As example, under 'statement of need' it switches from explaining the need directly to explaining how eyecite itself works.
Maybe this could be its own section?
