### 2021-09-20 #3617 @step21
> 👋 @step21
>
> > However there are tests in the repo, so I checked 'automated tests', as they can be automated.
>
> It's fine that you checked this, but it shouldn't be because they can be automated, but it can be simply because manual steps exist, as the criterion says.
>
> Thanks overall for your comments. If you can make any things that you think need to be changed but that don't yet have issues, such as your comment about the paper, can you create an issue for this in https://github.com/freelawproject/eyecite and mention `openjournals/joss-reviews#3617` in it?

Ok, thanks for clarifying. I made an issue about the paper structure.
