### 2021-09-20 freelawproject/eyecite#90 @mlissner
It is in the readme, but yeah, it'd be good to move it over to the contributor guidelines file. I'll take a stab at that.
