### 2021-09-21 0a2f381
Improves paper structure re #98.
