### 2021-09-21 freelawproject/eyecite#98 @mlissner
@step21, I just merged #99, which attempts to address this. Are you a good person to decide if this issue is thus resolved?
