### 2021-09-21 freelawproject/eyecite#98 @step21
Seems great. Thanks! I'd say as I raised the issue, I am also a good person to decide it is solved.
