+++ .github/workflows/pdoc.yml
@@ -1,0 +1,40 @@
+name: Pdoc
+
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+        with:
+          python-version: "3.8"
+
+      - name: Install pdoc
+      run: |
+        python -m pip install --upgrade pip3
+
+      - name: Run pdoc
+        working-directory: ./api_documentation
+        run: |
+          pdoc --html ./../eyecite --output-dir . --force
+
+      - name: Run sed  # Removes insanely-long parameter definition from docs
+        working-directory: ./api_documentation
+        run: |
+          sed -i '' 's/AhocorasickTokenizer(.*\]))/AhocorasickTokenizer()/' ./eyecite/find.html
+          sed -i '' 's/AhocorasickTokenizer(.*\]))/AhocorasickTokenizer()/' ./eyecite/index.html
+
+      - name: Commit changes
+        run: |
+          git config --local user.email "$(git log --format='%ae' HEAD^!)"
+          git config --local user.name "$(git log --format='%an' HEAD^!)"
+          git remote add github "https://$GITHUB_ACTOR:$GITHUB_TOKEN@github.com/$GITHUB_REPOSITORY.git"
+          git pull github ${GITHUB_REF} --ff-only
+          git add .
+          if [ -z "$(git status --porcelain)" ]; then
+            exit 0
+          fi
+          git commit -m "feat(docs): Update API documentation."
+          git push github HEAD:${GITHUB_REF}
