### 2021-09-23 68220cb
fix(ci): Fixes syntax error and uses deploy action
