### 2021-09-23 ccb38ea
fix(ci): install pdoc, not pip3
