+++ .github/workflows/pdoc.yml
@@ -13,7 +13,7 @@ jobs:
 
       - name: Install pdoc
         run: |
-          python -m pip install --upgrade pip3
+          python -m pip install --upgrade pdoc
 
       - name: Run pdoc
         working-directory: ./api_documentation
