### 2021-09-23 4af26aa
fix(ci): install pdoc, not pip3
