+++ .github/workflows/pdoc.yml
@@ -13,7 +13,7 @@ jobs:
 
       - name: Install pdoc
         run: |
-          python -m pip install --upgrade pdoc
+          python -m pip install --upgrade pdoc3
 
       - name: Run pdoc
         working-directory: ./api_documentation
