+++ .github/workflows/pdoc.yml
@@ -11,6 +11,26 @@ jobs:
         with:
           python-version: "3.8"
 
+      - name: Install Poetry
+        uses: snok/install-poetry@v1.1.0
+        with:
+          virtualenvs-create: true
+          virtualenvs-in-project: true
+
+      - name: Install hyperscan
+        run: sudo apt-get install libhyperscan-dev
+
+      - name: Load cached venv
+        id: cached-poetry-dependencies
+        uses: actions/cache@v2
+        with:
+          path: .venv
+          key: venv-${{ runner.os }}-${{ hashFiles('**/poetry.lock') }}-v2021-05-07
+
+      - name: Install dependencies
+        run: poetry install
+        if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+
       - name: Install pdoc
         run: |
           python -m pip install --upgrade pdoc3
