+++ .github/workflows/pdoc.yml
@@ -25,7 +25,7 @@ jobs:
         uses: actions/cache@v2
         with:
           path: .venv
-          key: venv-${{ runner.os }}-${{ hashFiles('**/poetry.lock') }}-v2021-05-07
+          key: venv-${{ runner.os }}-${{ hashFiles('**/poetry.lock') }}-v2021-09-23
 
       - name: Install dependencies
         run: poetry install
