### 2021-09-23 d8ef4a4
fix(ci): Try adding pdoc to pyproject
