+++ pyproject.toml
@@ -47,6 +47,7 @@ pylint-json2html = "^0.3.0"
 hyperscan = ">= 0.1.5"
 exrex = "^0.10.5"
 roman = "^3.3"
+pdoc3 = "^0.10.0"
 
 [tool.black]
 include = '''.*\.pyi?$'''
