+++ .github/workflows/pdoc.yml
@@ -31,10 +31,6 @@ jobs:
         run: poetry install
         if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
 
-      - name: Install pdoc
-        run: |
-          python -m pip install --upgrade pdoc3
-
       - name: Run pdoc
         working-directory: ./api_documentation
         run: |
