### 2021-09-23 f47370e
fix(ci): Fix the sed commands
