### 2021-09-23 10a8bf3
fix(ci): Only update docs when we merge to main
