### 2021-09-23 b590ede
fix(docs): Remove html files and add note to script
