+++ api_documentation/generate_documentation.sh
@@ -1,4 +1,9 @@
 #!/bin/bash
+
+# Documentation is typically generated by Github Action, but can be manually
+# generated with this script. Note that to make it live, it must be placed in
+# the gh-pages branch.
+
 (pip3 install pdoc3);
 (pdoc --html $(pwd)/../eyecite --output-dir $(pwd) --force);
 (sed -i '' 's/AhocorasickTokenizer(.*\]))/AhocorasickTokenizer()/' $(pwd)/eyecite/find.html)  # Removes insanely-long parameter definition
