+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-version = "2.2.0"
+version = "2.3.0"
 authors = ["Free Law Project <info@free.law>"]
 classifiers = [
   "Development Status :: 5 - Production/Stable",
