### 2021-09-23 faff64f
Generate API docs (#96)
* Renames annotate() to annotate_citations().

* Renames cleaners.py to clean.py.

* Moves clean_text() to clean module.

* Don't expose dump_citations() on top eyecite module.

* Renames find_citations.py to find.py.

* feat(docs): Generate API documentation using pdoc.

* feat(actions): Attempts to automate pdoc creation.

* fix(ci): Fixes syntax error and uses deploy action

* fix(ci): install pdoc, not pip3

* fix(ci): install pdoc, not pip3

* fix(ci): Updates flake8

* fix(ci): Updates pdoc

* fix(ci): Bust caches

* fix(ci): Try adding pdoc to pyproject

* fix(ci
[…]
