+++ eyecite/__init__.py
@@ -1,12 +1,20 @@
-from .annotate import annotate
-from .find_citations import get_citations
+from .annotate import annotate_citations
+from .clean import clean_text
+from .find import get_citations
 from .resolve import resolve_citations
-from .utils import clean_text, dump_citations
 
 __all__ = [
-    "annotate",
+    "annotate_citations",
     "get_citations",
     "clean_text",
     "resolve_citations",
-    "dump_citations",
 ]
+
+# No need to create API documentation for these internal helper functions
+__pdoc__ = {
+    "annotate.SpanUpdater": False,
+    "helpers": False,
+    "regexes": False,
+    "test_factories": False,
+    "utils": False,
+}
