+++ CHANGES.md
@@ -15,6 +15,16 @@ Fixes:
 
 ## Current
 
+**2.3.0 - 2021-09-23**
+
+Features:
+ - Greatly improved documentation
+ - Autogenerated documentation
+
+Changes:
+ - This version lands one more iteration of the APIs to make them more consistent. Sorry. Hopefully this will be the last of its kind for a while. The need for these changes became obvious when we began generating documentation. The changes are all in name only, not in functionality. So: 1) the `annotate` function is renamed as `annotate_citations`; 2) The `find_citations` module has been renamed `find` (so, do `from eyecite.find import get_citations` instead of `from eyecite.find_citations import get_citations`); 3) The `cleaners` module is now named `clean`; and 4) The `clean_text` function has been moved from `utils` to `clean` (so, do `from eyecite.clean import clean_text` instead of `from eyecite.utils import clean_text`). 
+
+
 **2.2.0 - 2021-06-04**
 
 Features:
