+++ .github/workflows/pdoc.yml
@@ -1,8 +1,9 @@
 name: Pdoc
 
-push:
-  branches:
-    - main
+on:
+  push:
+    branches:
+      - main
 
 jobs:
   build:
