+++ .github/workflows/pdoc.yml
@@ -55,4 +55,4 @@ jobs:
         uses: JamesIves/github-pages-deploy-action@4.1.5
         with:
           branch: gh-pages
-          folder: api_documentation
+          folder: api_documentation/eyecite
