### 2021-09-23 9fde09b
fix(ci): Only update the API docs on release
