+++ README.rst
@@ -25,9 +25,17 @@ Since eyecite was factored out from the CourtListener codebase into a standalone
 Contributions & Support
 =======================
 
-Please see the issues list on Github for things we need, or start a conversation if you have questions or need support. 
+Please see the issues list on Github for things we need, or start a conversation if you have questions or need support.
 
 If you are fixing bugs or adding features, before you make your first contribution, we'll need a signed contributor license agreement. See the template in the root of the repo for how to get that taken care of.
+
+API
+===
+The API documentation is located here:
+
+https://freelawproject.github.io/eyecite/
+
+It is autogenerated whenever we release a new version. Unfortunately, for now we do not support old versions of the API documentation, but it can be browsed in the gh-pages branch if needed.
 
 
 Quickstart
@@ -510,7 +518,7 @@ Deployment
 1. Update version info in :code:`pyproject.toml`.
 
 For an automated deployment, tag the commit with vx.y.z, and push it to master.
-An automated deploy will be completed for you.
+An automated deploy and documentation update will be completed for you.
 
 For a manual deployment, run:
 
@@ -518,6 +526,7 @@ For a manual deployment, run:
 
     poetry publish --build
 
+You will probably also need to push new documentation files to the gh-pages branch.
 
 
 Testing
