+++ .github/workflows/pdoc.yml
@@ -2,8 +2,8 @@ name: Pdoc
 
 on:
   push:
-    branches:
-      - main
+      tags:
+        - v*.*.*
 
 jobs:
   build:
