### 2021-09-23 freelawproject/eyecite#92 @mlissner
OK, this is now live at: https://freelawproject.github.io/eyecite/, and will auto-generate via a Github Action whenever a new version is tagged and pushed.

Thank you for this request!
