### 2021-09-24 aa1a770
fix(readme): Restores development credits to CAP.
