### 2021-09-24 freelawproject/eyecite#94 @mlissner
I think this should be all set. Do you want to take a look and close this issue if resolved?
