### 2021-09-24 freelawproject/eyecite#95 @mlissner
Hopefully this should be clear now too. Could you take a look and hopefully it'll be better? Thank you for the helpful feedback.
