### 2021-09-24 freelawproject/eyecite#95 @cmaimone
It's clearer, but I still think the code in the tutorial should be mostly run-able if possible.  For the custom_resolution_callback function, could you make a very simple one that just looked up values in a dict or something similar?  And also code to make your SomeCustomObject class?  This "class" gets used later for annotation examples too.

Or can you call one of the APIs if they're open (no key) to show a real example?

I'm not going to hold up the paper publication over this, but sorting out some of these things should help eyecite be more clearly and functionally a stand-alone package, that doesn't rely on such close integration with the systems it was developed for.
