### 2021-09-28 freelawproject/eyecite#95 @mattdahl
Hi @cmaimone, we've now extracted the tutorial into its own Jupyter notebook: https://github.com/freelawproject/eyecite/blob/main/TUTORIAL.ipynb. We think that this (1) provides a more realistic demonstration for the reader (e.g., by pulling a sample opinion from an API, as you suggested) and (2) makes clear that all the code is run-able by the reader (and modifiable by them, if they want to play around more with the example in Jupyter). What do you think?
