### 2021-09-28 freelawproject/eyecite#86 @mattdahl
Hi @step21, Jack added some notes about Hyperscan support here -- https://github.com/freelawproject/eyecite#hyperscan-installation. Does that resolve this issue?
