### 2021-09-28 79e6761
Adds percentage of recognized citations to paper.
