### 2021-09-29 #3617 @mattdahl
> I don't think I have any remaining issues that would prevent publication (a few open, but some can be closed, and others are for longer term consideration). Good work @mattdahl and team! I hope a lot of legal researchers are able to put this to good use.

Thanks @cmaimone for volunteering to review and for your careful attention. The project is much improved because of your suggestions!
