### 2021-10-02 #3617 @danielskatz
@mattdahl - I've now proofread the paper, and everything looks good to me.  You might want to check it again to make sure you are happy with it before we move to acceptance and publication.

The next steps are for you to:
