### 2021-10-07 #3617 @danielskatz
@mattdahl - We expect the archival deposit to be of the full repository, not just selected files.  For example, the tutorial is not in the repository.  The easiest way to do this is to clone the repository somewhere, then tar and gzip it, or just deposit the full cloned repository as a set of files in directories.
