### 2021-10-08 #3617 @danielskatz
> Just FYI, if you click on the "tree" view instead of the "table" view, you'll see that the directory structure is preserved. The relative file paths are also visible below the title of each file in the "table" view.

great!
